+++ NEWS.md
@@ -1,3 +1,12 @@
+# 1.6.0
+
+- New LabbcatEdit function
+  + *saveMedia* - saves the given media for the given transcript.
+  + *saveEpisodeDocument* - adds the given document to the given transcript's episode documents.
+  + *deleteMedia* - deletes media or episode document files.
+- Changed LabbcatEdit function
+  + GraphStore.newTranscript: rename mediaSuffix paramater as trackSuffix.
+  
 # 1.5.0
 
 - New LabbcatEdit functions